theme: Add insert-image partial and modify people-range to use it
Since we cannot use shortcodes in layout templates I am using a partial as an 'insert-image' function.
It will take two arguments: a target image, and a fallback image.
If the target image is not found it will insert the fallback image.
If neither is found, it will default to a 404_error image specified in the insert-image.html template.

This allows us to upload a fallback 404_error image and if other images are missing the site will not break.
(And more importantly, we might notice the missing assets sooner, while still allowing HUGO to build the site).