theme: Add person-cover class to scss, and implement {section}-cover class logic in article-cover partial
Decided to use a "person-cover" class div inside the article-cover div.
This class div is actually named according to the section of the current page.
If the page is a single page, as opposed to a section list page, the class is
made singular, i.e: people --> person

NOTE: Keep in mind that this only works for sections or pages under sections.
The 'About' page, for instance, which is a single page under the root/home dir,
and not under any section does not correctly populate the above class.